Changed addresses to unsigned
Finally ripped the band-aid off and changed all register address variables to unsigned. Was getting too hard to worry about sign extensions and make silly casts.
Doing so revealed one bug - WriteBytes() was trying to use SetNull() to clear the unused part of the buffer, but always calculated a negative length, doing nothing instead.
Also changed many lengths to unsigned where they clearly were, while leaving offsets signed if there's a possibility they could be negative.
Also renamed the IntBE functions to UIntBE where that was already the case, and generally made things more consistent.
Note there is one change left: SetBit(), but I couldn't tease it out of other changes so will be in a subsequent commit.

+++ src/Streams/GvspAck.cs
@@ -115,9 +115,9 @@ namespace GigE_Cam_Simulator.Streams
             PacketFormat = format;
         }
 
-        public BufferReader CreateBuffer(int length)
+        public BufferReader CreateBuffer(uint length)
         {
-            var b = new BufferReader(length + 4 * 2);
+            var b = new BufferReader((int)length + 4 * 2);
 
             b.WriteWordBE(Status);
             b.WriteWordBE(BlockId);
